Refactor notification timer with useNotificationTimer and replace css of UserForm
