+++ frontend/src/components/UserForm/UserForm.css
@@ -1,14 +1,14 @@
-.login-form {
+.user-form {
     max-width: 350px;
     margin: auto;
 }
 
-.login-form-submit {
+.user-form-submit {
     margin-top: 50px;
 }
 
-.login-form-submit:disabled,
-.login-form-submit[disabled]{
+.user-form-submit:disabled,
+.user-form-submit[disabled]{
     border: 1px solid #999999;
     background-color: #cccccc;
     color: #666666;
