+++ frontend/src/pages/FlightPage/components/FlightList/FlightList.tsx
@@ -5,10 +5,11 @@ import {calculateDuration, formatDate, formatTime} from "../../../../utils/funct
 import {Link, useNavigate} from "react-router-dom";
 import axios from "axios";
 import ConfirmationModal from "../../../../components/ConfirmationModal/ConfirmationModal.tsx";
-import {useEffect, useState} from "react";
+import {useState} from "react";
 import DeleteIcon from '@mui/icons-material/Delete';
 import Notification from "../../../../components/Notification/Notification.tsx";
 import {useFetchOptions} from "../../../../hooks/useFetchOptions.ts";
+import {useNotificationTimer} from "../../../../hooks/useNotificationTimer.ts";
 
 type FlightListProps = {
     data: Flight[],
@@ -48,15 +49,7 @@ export default function FlightList({ data, fetchAllFlights }: Readonly<FlightLis
         }
     }
 
-    useEffect(() => {
-        if (showNotification) {
-            const timer = setTimeout(() => {
-                setShowNotification(false);
-            }, 5000);
-
-            return () => clearTimeout(timer);
-        }
-    }, [showNotification]);
+    useNotificationTimer(showNotification, setShowNotification);
 
     return(
         <div>
