+++ frontend/src/pages/LoginPage/LoginPage.tsx
@@ -1,6 +1,5 @@
-import './LoginPage.css';
 import {UserForLogin} from "../../types/auth/userType.ts";
-import {ChangeEvent, Dispatch, FormEvent, SetStateAction, useEffect, useState} from "react";
+import {ChangeEvent, Dispatch, FormEvent, SetStateAction, useState} from "react";
 import UserForm from "../../components/UserForm/UserForm.tsx";
 
 type LoginPageProps = {
@@ -26,21 +25,12 @@ export default function LoginPage({ login, setShowLoginNotification, showLoginNo
         login(user);
     }
 
-    useEffect(() => {
-        if (showLoginNotification) {
-            const timer = setTimeout(() => {
-                setShowLoginNotification(false);
-            }, 5000);
-
-            return () => clearTimeout(timer);
-        }
-    }, [showLoginNotification]);
-
     return (
         <article>
             <UserForm
                 showNotification={showLoginNotification}
                 setShowNotification={setShowLoginNotification}
+                notificationMessage={"The username and/or password are not found. Please check again."}
                 pageName={"Login"}
                 formType={"login"}
                 handleSubmit={handleLogin}
